refactor: organizing assets folder
