+++ frontend/src/assets/images/purple_background.png
