+++ frontend/src/assets/images/purple_row.png
