+++ frontend/src/pages/home/components/header/HomeHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import {ReactComponent as BitwiseLogo} from '../../../../assets/bitwise_logo.svg';
+import {ReactComponent as BitwiseLogo} from '../../../../assets/icons/bitwise_logo.svg';
 
 import {StyledHeader} from './HomeHeaderStyles';
