+++ frontend/src/pages/home/components/search_container/HomeSearchContainerStyles.ts
@@ -1,7 +1,7 @@
 
 import styled from 'styled-components';
 
-import {ReactComponent as SearchIcon} from '../../../../assets/search_icon.svg';
+import {ReactComponent as SearchIcon} from '../../../../assets/icons/search_icon.svg';
 
 
 export const StyledSearchContainer = styled.section`
