+++ frontend/src/pages/search/components/footer/SearchFooter.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { ReactComponent as BitwiseLogo } from "../../../../assets/bitwise_logo_2.svg";
+import { ReactComponent as BitwiseLogo } from "../../../../assets/icons/bitwise_logo_2.svg";
 
 import SocialMedia from "../../../../components/social_media/SocialMedia";
